Division operation used.
Hi Tushar
Firstly, thank you for the repository and solutions for all the questions. It's really helpful to have all great questions at one place.
I had a question in this solution.
Line number 43 and 44 are using division operation. I was under the assumption that we can't use it for this question. Please clarify the approach/question.